2.1 Divided main.js into modules
