+++ public/index.html
@@ -13,15 +13,7 @@
     <h1>Предназначение дня</h1>
   </header>
   <main>
-    <div class="days">
-      <!-- <div class="days__item" data-day="mon">Понедельник</div>
-      <div class="days__item" data-day="tue">Вторник</div>
-      <div class="days__item" data-day="wed">Среда</div>
-      <div class="days__item" data-day="thu">Четверг</div>
-      <div class="days__item" data-day="fri">Пятница</div>
-      <div class="days__item" data-day="sat">Суббота</div>
-      <div class="days__item" data-day="sun">Воскресенье</div> -->
-    </div>
+    <div class="days"></div>
   </main>
 
   <script src="bundle.js"></script>
