+++ public/styles/style.scss
@@ -39,6 +39,6 @@ main {
   &__item {
     width: 14.3%;
     max-height: 100vh;
-    // border: 1px solid black;
+    border: 1px solid black;
   }
 }
