+++ public/test.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="styles/style.css">
+  <!-- <link rel="shortcut icon" href="favicon.ico" type="image/x-icon"> -->
+  <title>Mission of the day</title>
+</head>
+<body>
+  
+  <header>
+    <h1>Предназначение дня</h1>
+  </header>
+  <main>
+    <div class="days">
+      <div class="days__item" data-day="mon">Понедельник</div>
+      <div class="days__item" data-day="tue">Вторник</div>
+      <div class="days__item" data-day="wed">Среда</div>
+      <div class="days__item" data-day="thu">Четверг</div>
+      <div class="days__item" data-day="fri">Пятница</div>
+      <div class="days__item" data-day="sat">Суббота</div>
+      <div class="days__item" data-day="sun">Воскресенье</div>
+    </div>
+  </main>
+
+  <!-- <script src="bundle.js"></script> -->
+</body>
+</html>
