+++ src/utils/const.js
@@ -1,0 +1,7 @@
+const RenderPosition = {
+  AFTER_BEGIN: `afterbegin`,
+  AFTER_END: `afterend`,
+  BEFORE_END: `beforeend`,
+};
+
+export {RenderPosition};
